Update c2e5 template and answer
