+++ deb/ch2/ep1/answers/fruits.py
