+++ deb/ch2/ep1/answers/tollbooth_part2.py
