+++ deb/ch2/ep1/answers/tollbooth_part3.py
