+++ deb/ch2/ep1/answers/tollbooth_part4.py
