+++ deb/ch2/ep3/answers/part4.py
@@ -1,21 +1,172 @@
+"""
+Ingest flight scheduled from REST API into Google BigQuery using Cloud Dataflow (Apache Beam).
 
-import sys
-import os
-import csv
+
+Author: Par (turalabs.com)
+Contact:
+
+license: GPL v3 - PLEASE REFER TO DEB/LICENSE FILE
+"""
+
 import argparse
-from io import StringIO
-from datetime import datetime, time
-from time import time as now
+import json
+import os
+import sys
+import typing
+from copy import copy
+from datetime import datetime, timedelta
+from time import sleep, time
 
 import apache_beam as beam
-from apache_beam.options.pipeline_options import SetupOptions, PipelineOptions
+import requests
+from apache_beam.options.pipeline_options import PipelineOptions, SetupOptions
 
-from deb.utils.logging import logger
-from deb.utils.config import config
 from deb.utils.answers.bq_utils import BigQueryUtils
-from deb.utils.data_models.flights import (FLIGHTS_CSV_COLUMNS,
-                                           datamodel_flights_parquet_schema,
-                                           datamodel_flights_bigquery_schema)
+from deb.utils.config import config
+from deb.utils.logging import logger
+from deb.utils.data_models.flights import FUTURE_FLIGHTS_BIGQUERY_SCHEMA
+
+# module variables
+api_url = 'http://localhost:5000/'          # todo:: change with app-engine url
+api_token = '<YOUR OWN API TOKEN>'
+api_timeout = 30.0
+api_max_retries = 3
+
+# module public cached values
+AIRLINES = []
+
+
+# ====================================================================
+# ‹‹=‹‹=‹‹=‹‹=‹‹=‹‹=‹‹=‹‹= REST API FUNCTIONS =››=››=››=››=››=››=››=››
+# ====================================================================
+
+
+def api_get_airlines(api_url=api_url,
+                     api_token=api_token,
+                     tries=0,
+                     timeout=api_timeout) -> typing.List[str]:
+    """
+    Call the REST API to get the list of available airlines
+
+    :param api_url:
+    :param api_token:
+    :param tries:
+    :param timeout:
+    :param kwargs:
+    :return:
+    """
+    global AIRLINES
+    if not AIRLINES:
+        headers = {'Authorization': f'Bearer {api_token}',
+                   'Accept': 'application/json'}
+        url = os.path.join(api_url, 'flights/airlines')
+        logger.debug(f"Calling airlines REST API (url={url})")
+        try:
+            r = requests.get(url, headers=headers, timeout=timeout)
+            r.raise_for_status()
+            AIRLINES = r.json()['airlines']
+            logger.info(f"flight airlines api call success in {r.elapsed.total_seconds():.3}s. Airlines: {AIRLINES}")
+            return AIRLINES
+        except requests.exceptions.ConnectTimeout:
+            # REST API connection timeout is reached. Most likely we're sending too many requests and need to ease off
+            # let's exponentially increase the timeout and retry until the max num of retires is reached
+            logger.warning(f"REST API call to get airlines timed out. (timeout={timeout}")
+            tries, timeout = tries + 1, timeout * 2
+            if tries < api_max_retries:
+                # sleep the current thread to back off from submitting too many requests
+                logger.debug(f"thread sleep to back off from submitting too many REST API requests. sleep: {timeout}s")
+                sleep(timeout)
+                logger.warning(f"Increasing timeout to {timeout} and retrying. Retry number={tries}")
+                return api_get_airlines(api_url, api_token, tries, timeout)
+            else:
+                logger.fatal("Max number of API retries is reached. Quiting.")
+                sys.exit(1)
+        except requests.ConnectionError:
+            # if cannot establish connection with the REST API. Most likely the URL is incorrect or API Flask server
+            # is not running
+            logger.fatal(f"Could not establish connection to the REST API to get airlines (url={url})")
+            sys.exit(1)
+        except requests.exceptions.RequestException as err:
+            logger.error("Unknown error while connecting to REST API to get airlines: {}".format(str(err)))
+            return AIRLINES
+    else:
+        return AIRLINES
+
+
+def api_get_flights(airline: str,
+                    flight_date: datetime,
+                    api_url=api_url,
+                    api_token=api_token,
+                    tries=0,
+                    timeout=api_timeout) -> typing.List[typing.Dict]:
+    """
+    Call REST API to get flight records for a given date
+
+    :param airline: airline IATA code
+    :param flight_date: flight date
+    :param api_url: API base URL such as http://localhost:5000/
+    :param api_token: OAuth Bearer token
+    :param tries:
+    :param timeout:
+    :return: list of flight records as a dict
+    """
+    headers = {'Authorization': f'Bearer {api_token}',
+               'Accept': 'application/json'}
+    params = {'date': flight_date.strftime('%Y-%m-%d'),
+              'airline': airline}
+    url = os.path.join(api_url, 'flights/flights')
+    try:
+        r = requests.get(url, params=params, headers=headers, timeout=timeout)
+        # raise exception if response returned an unsuccessful status code
+        r.raise_for_status()
+        # return flight records
+        data = r.json()
+        logger.debug(f"Flight API returned {data['records']} flights in {r.elapsed.total_seconds():.3}s")
+        return copy(data['flights'])
+    except requests.exceptions.ConnectTimeout:
+        # REST API connection timeout is reached. Most likely we're sending too many requests and need to ease off
+        # let's exponentially increase the timeout and retry until the max num of retires is reached
+        logger.warning(f"REST API call to get flights timed out. (timeout={timeout}")
+        tries, timeout = tries + 1, timeout * 2
+        if tries < api_max_retries:
+            # sleep the current thread to back off from submitting too many requests
+            logger.debug(f"thread sleep to back off from submitting too many REST API requests. sleep: {timeout}s")
+            sleep(timeout)
+            logger.warning(f"Increasing timeout to {timeout} and retrying. Retry number={tries}")
+            return api_get_flights(airline, flight_date, api_url, api_token, tries, timeout)
+        else:
+            logger.fatal("Max number of API retries is reached. Quiting.")
+            sys.exit(1)
+    except requests.ConnectionError:
+        # if cannot establish connection with the REST API. Most likely the URL is incorrect or API Flask server
+        # is not running
+        logger.fatal(f"Could not establish connection to the REST API to get flights (url={url})")
+        sys.exit(1)
+    except requests.exceptions.HTTPError as err:
+        logger.error("Flights API has a critical error: {}".format(str(err)))
+        logger.error("No records returned by flights api")
+        return []
+    except requests.exceptions.RequestException as err:
+        logger.error("Unknown error while connecting to REST API to get flights: {}".format(str(err)))
+        logger.error("No records returned by flights api")
+        return []
+
+
+# ==================================================================
+# ‹‹=‹‹=‹‹=‹‹=‹‹=‹‹=‹‹=‹‹= HELPER FUNCTIONS =››=››=››=››=››=››=››=››
+# ==================================================================
+
+
+def list_dates(start_date: datetime, end_date: datetime) -> typing.List[datetime]:
+    """
+    Create a list of dates between two dates including both endpoint dates
+
+    :param start_date: starting date
+    :param end_date: ending date
+    :return: Array. List of dates in between start_date and end_date
+    """
+    num_days = (end_date - start_date).days + 1        # add 1 to include end date
+    return [start_date + timedelta(days=x) for x in range(num_days)]
 
 
 # =================================================================
@@ -23,300 +174,227 @@ from deb.utils.data_models.flights import (FLIGHTS_CSV_COLUMNS,
 # =================================================================
 
 
-class BeamReadCSV(beam.DoFn):
-
-    def __init__(self, header_cols=FLIGHTS_CSV_COLUMNS):
-        self.cols = header_cols
-        self.row_count = 0
-        self.bad_rows = 0
-        super(BeamReadCSV, self).__init__()
-
-    def setup(self):
-        self.row_count = 0
-        self.bad_rows = 0
-
-    def teardown(self):
-        logger.info(f"total rows: {self.row_count}, bad row: {self.bad_rows}")
-
-    def process(self, element, *args, **kwargs):
-        """
-        Parse a records CSV line and transpose column headers
-
-        :param element: csv line
-        :return: dict {column_name: csv_value}
-        """
-        # use csv.reader to correctly parse csv. accounting for quoted values
-        reader = csv.reader(StringIO(element), delimiter=',')
-        for row in reader:
-            if len(row) == len(self.cols):
-                # transpose the schema onto the csv cols to create a
-                # {column_name: csv_value} dict. read about python zip function: https://www.programiz.com/python-programming/methods/built-in/zip
-                d = dict(zip(self.cols, row))
-                self.row_count += 1
-                yield d
-            else:
-                self.bad_rows += 1
-                logger.debug(f"bad row: {row}")
-
-
-class BeamTransformRecords(beam.DoFn):
-
-    def process(self, element: dict, *args, **kwargs):
-        try:
-            date_fmt = kwargs['date_fmt'] if 'date_fmt' in kwargs else '%Y-%m-%d'
-            time_fmt = kwargs['time_fmt'] if 'time_fmt' in kwargs else '%H%M'
-
-            # convert midnight time: '2400' to '0000'. causes `datetime.strptime()` to throw ValueError exception
-            time_cols = ['departure_time', 'actual_departure_time', 'wheels_off', 'wheels_on', 'arrival_time', 'actual_arrival_time']
-            for col in time_cols:
-                if element[col] == '2400':
-                    element[col] = '0000'
-
-            # parse records
-            # Pay attention how we use python inline `<expression_if_true> if <condition> else <expression_if_false>` to assert default values
-            # for some columns. Parsing columns which are mandatory do not have defaults and will result in an exception
-            element['flight_date'] = datetime.strptime(element['flight_date'], date_fmt).date()
-            element['day_of_week'] = int(element['day_of_week']) if element['day_of_week'] else element['flight_date'].weekday() + 1
-            # remove state from src_city
-            element['src_city'] = str(element['src_city']).split(',')[0]
-            element['departure_time'] = datetime.strptime(element['departure_time'], time_fmt).time()
-            element['actual_departure_time'] = datetime.strptime(element['actual_departure_time'], time_fmt).time() if element['actual_departure_time'] else None
-            element['departure_delay'] = float(element['departure_delay']) if element['departure_delay'] else 0.0
-            element['taxi_out'] = float(element['taxi_out']) if element['taxi_out'] else None
-            element['wheels_off'] = datetime.strptime(element['wheels_off'], time_fmt).time() if element['wheels_off'] else None
-            element['wheels_on'] = datetime.strptime(element['wheels_on'], time_fmt).time() if element['wheels_on'] else None
-            element['taxi_in'] = float(element['taxi_in']) if element['taxi_in'] else None
-            element['arrival_time'] = datetime.strptime(element['arrival_time'], time_fmt).time()
-            element['actual_arrival_time'] = datetime.strptime(element['actual_arrival_time'], time_fmt).time() if element['actual_arrival_time'] else None
-            element['arrival_delay'] = float(element['arrival_delay']) if element['arrival_delay'] else 0.0
-            element['cancelled'] = str(element['cancelled']).strip() == '1'
-            element['cancellation_code'] = element['cancellation_code'] if element['cancellation_code'] else None
-            element['flight_time'] = float(element['flight_time']) if element['flight_time'] else None
-            element['actual_flight_time'] = float(element['actual_flight_time']) if element['actual_flight_time'] else None
-            element['air_time'] = float(element['air_time']) if element['air_time'] else None
-            element['flights'] = int(element['flights']) if element['flights'] else 1
-            element['distance'] = float(element['distance'])
-            element['airline_delay'] = float(element['airline_delay']) if element['airline_delay'] else 0.0
-            element['weather_delay'] = float(element['weather_delay']) if element['weather_delay'] else 0.0
-            element['nas_delay'] = float(element['nas_delay']) if element['nas_delay'] else 0.0
-            element['security_delay'] = float(element['security_delay']) if element['security_delay'] else 0.0
-            element['late_aircraft_delay'] = float(element['late_aircraft_delay']) if element['late_aircraft_delay'] else 0.0
-
-            # transform time field to HH:MM:SS format. BigQuery friendly
-            for k in time_cols:
-                if element[k] is not None and isinstance(element[k], time):
-                    element[k] = element[k].strftime('%H:%M:%S')
-
-            yield element
-        except (KeyError, ValueError, TypeError) as err:
-            # logger.debug(f"invalid record: {element}")
-            yield beam.pvalue.TaggedOutput('rejects', element)
-
-
-class BeamLookupRoute(beam.DoFn):
+class BeamExpandDaysByAirlines(beam.DoFn):
     """
-    Takes list of airports as a pandas dataframe airport and looks up src and dest airport city, state.
-    It also produces a secondary output with list of airports which are missing in the look up dataframe.
+    Expands a date with all possible airline combinations. Takes in a date and will produce multiple
+    output records with each airline and the date
     """
 
-    def process(self, element, routes, *args, **kwargs):
-        """
-        Lookup the flight route (airline, src, dest) in routes lookup table.
+    def process(self, element, airlines=[], *args, **kwargs) -> typing.List[typing.Tuple]:
+        return [(element, _) for _ in airlines]
 
-        Reject the flight if not found in routes (as 'rejects' output PCollection) AND output missing routes as
-        'missing_routes' output PCollection.
 
-        :param element: input flight record as dict
-        :param routes: routes lookup table with ('airline', 'src', 'dest') tuple as keys
-        :return: output flight dict
-        """
+class BeamGetFlights(beam.DoFn):
+    """
+    Call the Flights API and return scheduled flight records. Takes in a tuple of (airline, flight_date) and
+    returns flight records from calling the api
+    """
+
+    def process(self, element: typing.Tuple, *args, **kwargs) -> typing.List[typing.Dict]:
+        fligh_date, airline = element
+        flights = api_get_flights(airline, fligh_date, api_url=kwargs['api_url'], api_token=kwargs['api_token'])
+        logger.info(f"flights api call airline={airline}, flight_date={fligh_date} returned {len(flights)} flights")
+        return flights
+
+
+class BeamTransformFlights(beam.DoFn):
+    """
+    Transform flights records. Parse dates and times and add day_of_week field
+    """
+
+    def process(self, element: typing.Dict, *args, **kwargs) -> typing.List[typing.Dict]:
         try:
-            # route key as (airline, src, dest) tuple
-            route = (element['airline'], element['src'], element['dest'])
-            if route not in routes:
-                # missing route. reject the flight record and add as missing
-                yield beam.pvalue.TaggedOutput('rejects', element)  # send the record to rejects
-                yield beam.pvalue.TaggedOutput('missing_routes', (route, 1))  # also send the route as missing
-            # regardless pass the record through to the main output
-            # change this to an else statement if you wish NOT to pass the element through if missing route
+            # get date format and time format from optional function arguments
+            date_fmt = kwargs["date_format"] if "date_format" in kwargs else '%Y-%m-%d'  # set date format to YYYY-MM-DD by default or take it from args
+            time_fmt = kwargs["time_format"] if "time_format" in kwargs else '%H%M'  # set time format to HHMM by default or take it from args
+            # parse flight dates and times and add day of the week
+            flight_date = datetime.strptime(element['flight_date'], date_fmt).date()
+            element['flight_date'] = flight_date.strftime('%Y-%m-%d')  # bigquery friendly formatted date
+            element['departure_time'] = datetime.strptime(element['departure_time'], time_fmt).strftime('%H:%M:%S')  # bigquery friendly formatted time
+            element['arrival_time'] = datetime.strptime(element['arrival_time'], time_fmt).strftime('%H:%M:%S')  # bigquery friendly formatted time
+            element['day_of_week'] = flight_date.weekday() + 1      # add 1 to start Monday as 1, ending with Sunday as 7
             yield element
-        except (KeyError, ValueError, TypeError) as err:
-            logger.debug(f"input flight record is missing critical fields or of invalid type.")
+        except TypeError as err:
+            logger.warning(f"input flight record is not a proper dict. omitting output")
+        except (KeyError, ValueError) as err:
+            logger.debug(f"intput flight record is missing critical fields. omitting output")
             logger.debug(str(err))
+
 
 # ===============================================================
 # ‹‹=‹‹=‹‹=‹‹=‹‹=‹‹=‹‹=‹‹= RUN FUNCTIONS =››=››=››=››=››=››=››=››
 # ===============================================================
 
-
 def runtime_args(args=sys.argv):
-    """
-    Parse command line arguments. Type --help to see list of args.
+    """Parse runtime args"""
 
-    :param args: sys.argv
-    :return: (known args, beam pipeline args)
-    """
+    def parse_date(date_str):
+        try:
+            return datetime.strptime(date_str, '%Y-%m-%d')
+        except ValueError:
+            raise argparse.ArgumentTypeError(f"Invalid date format ('{date_str}'). Date must be in YYYY-MM-DD format.")
 
-    # set default values from config.yaml file
-    # *** NOTE ***:
-    #   if you get a confuse.exceptions.NotFoundError exception:
-    #   add a default value under defaults.chXX.epYY in config.yaml
-    defaults = config['defaults']['ch2']['ep3']     # default config path for this episode
-    default_input = defaults['input'].as_str()
+    defaults = config['defaults']['ch2']['ep5']
+    default_start_date = parse_date(defaults['start_date'].as_str())
+    default_end_date = parse_date(defaults['end_date'].as_str())
+    # api options
+    global api_url, api_token, api_timeout, api_max_retries
+    api_url = defaults['api_url'].as_str()
+    api_token = defaults['api_token'].as_str()
+    api_timeout = defaults['api_timeout'].get(float)
+    api_max_retries = defaults['api_max_retries'].get(int)
+    # cloud Options
     default_output = defaults['output'].as_str()
     default_flights_ext_table = defaults['flights_ext_table'].as_str()
     default_flights_table = defaults['flights_table'].as_str()
+    default_airports_table = defaults['airports_table'].as_str()
     default_routes_table = defaults['routes_table'].as_str()
 
-    p = argparse.ArgumentParser(description='Apache Beam (Google Dataflow) process to ingest historical flight records',
+    p = argparse.ArgumentParser(description='Apache Beam (Google Dataflow) process to query flight schedule records via'
+                                            'REST API and ingest into Google BigQuery.',
                                 formatter_class=argparse.RawTextHelpFormatter)
-    p.add_argument('--input', default=default_input, required=False, type=str,
-                   help=('Input path. Local or gcs URI such as: \n'
-                         'gs://deb-airline-data/bots/csv/2018/*.csv'))
+    p.add_argument('--start-date', default=default_start_date, required=False, type=parse_date,
+                   help=('Starting date wto fetch flights records as: \n'
+                         'YYYY-MM-DD ie: 2020-01-15'))
+    p.add_argument('--end-date', default=default_end_date, required=False, type=parse_date,
+                   help=('End date wto fetch flights records as: \n'
+                         'YYYY-MM-DD ie: 2020-01-15'))
     p.add_argument('--output', default=default_output, required=False, type=str,
-                   help=('Input path. Local or gcs URI such as: \n'
-                         'gs://deb-airline-data-etl/beam/ch2cp3/'))
+                   help=('Google Storage output bucket path for the external table. ie: \n'
+                         'gs://bucket-name/beam/ch2ep5/\n\n'
+                         'Be sure to include the trailing / in path.\n\n'
+                         'Warning: bucket location must be the same BigQuery dataset location\n'
+                         'Warning: content of this bucket is emptied upon execution\n'))
     p.add_argument('--flights-ext-table', default=default_flights_ext_table, required=False, type=str,
                    help=('BigQuery external table name for staging flights records as: \n'
                          'dataset_name.table.name'))
     p.add_argument('--flights-table', default=default_flights_table, required=False, type=str,
                    help=('BigQuery final flights output table name:\n'
                          'dataset_name.table.name'))
+    p.add_argument('--airports-table', default=default_airports_table, required=False, type=str,
+                   help=('BigQuery airports lookup table name:\n'
+                         'dataset_name.table.name'))
     p.add_argument('--routes-table', default=default_routes_table, required=False, type=str,
                    help=('BigQuery routes lookup table name:\n'
                          'dataset_name.table.name'))
-
+    p.add_argument('--api-url', default=api_url, required=False, type=str,
+                   help=('Flight API full URL as:\n'
+                         'http://localhost:5000/\n\n'
+                         'Be sure to include the trailing /\n'))
+    p.add_argument('--api-token', default=api_token, required=False,
+                   help=('Flights API Bearer auth token'))
     # parse args
     known_args, beam_args = p.parse_known_args(args)
-    # push args to config, confuse adds args to the top level of yaml config
-    # parameters would be accessible as config['param_name']
+    # set command line args into config. confuse will load args at the top level
     config.set_args(known_args)
 
-    # print arguments
-    logger.info("arguments:")
+    # print args
+    logger.info("printing args:")
     for k, v in known_args.__dict__.items():
         logger.info(f"\t{k}={v}")
-    logger.info(f"beam args: {beam_args}")
+    logger.info(f"pipeline args: {beam_args}")
 
-    # return both known command line args and apache beam args
+    # return args
     return known_args, beam_args
 
 
-def run():
-    t0 = now()
+def run_simple():
+    t0 = time()
 
-    # parse command line options
+    # parse command line arguments
     known_args, beam_args = runtime_args()
 
-    # BigQuery utility
+    # BigQuery Utility
     bq_utils = BigQueryUtils()
 
+    # pass in the pipeline options
     options = PipelineOptions(beam_args)
     options.view_as(SetupOptions).save_main_session = True
-
     with beam.Pipeline(options=options) as p:
-        rows = (p
-                | beam.io.ReadFromText(known_args.input, skip_header_lines=1)
-                | beam.ParDo(BeamReadCSV(header_cols=FLIGHTS_CSV_COLUMNS))
-                | beam.ParDo(BeamTransformRecords(), date_fmt='%Y-%m-%d', time_fmt='%H%M')
-                )
+        # pre-process: create a list of date to process and get other side-inputs
+        # create a list of flights date to retrieve from api
+        days = list_dates(start_date=known_args.start_date, end_date=known_args.end_date)
 
-        # load the routes table into a lookup dict
-        sql = f"""select airline, src, dest from {known_args.routes_table}"""
-        routes = bq_utils.execute_as_dict(sql, keycols=['airline', 'src', 'dest'])
+        # get airline iata codes from the api
+        airlines = api_get_airlines(api_url=known_args.api_url, api_token=known_args.api_token)
 
-        # lookup routes
-        rows, routes_rejects, missing_routes = (rows
-                                                | beam.ParDo(BeamLookupRoute(), routes=routes).with_outputs('rejects', 'missing_routes', main='main'))
+        # create a beam collection with all days and airlines to get flights for
+        input_rows = (p
+                      | beam.Create(days)
+                      | beam.ParDo(BeamExpandDaysByAirlines(), airlines=airlines)
+                      )
 
-        # write parquet output files
-        output = (rows
-                  | beam.io.WriteToParquet(os.path.join(known_args.output, 'flights'),
-                                           schema=datamodel_flights_parquet_schema(),
-                                           file_name_suffix='.parquet')
-                  )
+        # call flights api to get flights for each record above and
+        # call the beam transforms to process the input flights
+        flights = (input_rows
+                   | beam.ParDo(BeamGetFlights(), api_url=known_args.api_url, api_token=known_args.api_token)
+                   | beam.ParDo(BeamTransformFlights())
+                   )
 
-        # write missing routes to another output as CSV
-        output_routes = (missing_routes
-                         | "gbr" >> beam.GroupByKey()   # calculate distinct missing routes
-                         | "missing_routes_csv" >> beam.Map(lambda e: ','.join(list(e[0]))) # csv output the key (e[0] of key value tuple) which is (airline,src,dest)
-                         | "missing_routes_out" >> beam.io.WriteToText(os.path.join(known_args.output, 'rejects/missing-routes'),
-                                                                       file_name_suffix='.csv',
-                                                                       header='airline,src,dest')
-                         )
+        # prepare & write output files
+        json_output = (flights
+                       | beam.Map(lambda e: json.dumps(e))
+                       | beam.io.WriteToText(os.path.join(known_args.output, 'flights'), file_name_suffix='.json')
+                       )
 
-        # alternative: write (simple) newline delimited json output files
-        #              a very flexible output file format for bigquery and other big data tools
-        #              much slower to write and larger in size than binary formats such as Parquet, ORC, or Avro
-        #              but provides flexibility over schema for smaller data files
-        #              larger file sizes should use Avro, Parquet, ORC. Avro provides fastest write speeds where
-        #              parquet and orc provide faster read performance for analytical queries
+    logger.info("apache beam pipeline done")
 
-        # output = (rows
-        #           | beam.Map(lambda e: {k: v if k != 'flight_date' else v.strftime('%Y-%m-%d') for k, v in e.items()})  # convert flight_date back to string type for json conversion
-        #           | beam.Map(lambda e: json.dumps(e))  # json dump row
-        #           | beam.io.WriteToText(os.path.join(known_args.output, 'flights'),
-        #                                 file_name_suffix='.json')
-        #           )
-
-    logger.info("beam pipiline completed.")
-
-    # create bigquery external table and insert into bq flights table
+    # create bigquery external table
+    logger.info("dropping and creating bigquery external table...")
     bq_utils.create_external_table(known_args.flights_ext_table,
-                                   source_uris=os.path.join(known_args.output, "flights*.parquet"),
-                                   source_format='PARQUET',
+                                   source_uris=os.path.join(known_args.output, "flights*.json"),
+                                   schema=FUTURE_FLIGHTS_BIGQUERY_SCHEMA,
                                    delete_if_exists=True)
 
-    # create and replace existing bigquery flights table
-    bq_utils.create_table(known_args.flights_table, schema=datamodel_flights_bigquery_schema(), delete_if_exists=True)
+    # delete existing flights greater than start_date
+    sql = f"""DELETE FROM {known_args.flights_table} WHERE flight_date >= '{datetime.strftime(known_args.start_date, '%Y-%m-%d')}'"""
+    bq_utils.execute(sql)
 
-    # insert into table as select (itas) statement
-    sql = f"""
-        INSERT INTO `{known_args.flights_table}`
-        SELECT
-          a.day_of_week,
-          a.flight_date,
-          a.airline,
-          a.tailnumber,
-          a.flight_number,
-          a.src,
-          a.src_city,
-          a.src_state,
-          a.dest,
-          a.dest_city,
-          a.dest_state,
-          PARSE_TIME('%H:%M:%S', a.departure_time) as departure_time,
-          PARSE_TIME('%H:%M:%S', a.actual_departure_time) as actual_departure_time,
-          a.departure_delay,
-          a.taxi_out,
-          PARSE_TIME('%H:%M:%S', a.wheels_off) as wheels_off,
-          PARSE_TIME('%H:%M:%S', a.wheels_on) as wheels_on,
-          a.taxi_in,
-          PARSE_TIME('%H:%M:%S', a.arrival_time) as arrival_time,
-          PARSE_TIME('%H:%M:%S', a.actual_arrival_time) as actual_arrival_time,
-          a.arrival_delay,
-          a.cancelled,
-          a.cancellation_code,
-          a.flight_time,
-          a.actual_flight_time,
-          a.air_time,
-          a.flights,
-          a.distance,
-          a.airline_delay,
-          a.weather_delay,
-          a.nas_delay,
-          a.security_delay,
-          a.late_aircraft_delay,
-          -- CONCAT(a.flight_date, '_', a.airline, '_', a.flight_number) AS flightDate_airline_flightNumber
-        FROM
-          `{known_args.flights_ext_table}`  a
-        """
-    # insert records form parquet external table into final bq managed flights table
-    r = bq_utils.execute(sql)
+    # insert flights records into final table
+    sql = f"""INSERT INTO {known_args.flights_table}
+                        (SELECT
+                          a.day_of_week ,
+                          a.flight_date ,
+                          a.airline ,
+                          a.tailnumber ,
+                          a.flight_number ,
+                          a.src ,
+                          b.city as src_city ,
+                          b.state as src_state ,
+                          a.dest ,
+                          c.city as dest_city ,
+                          c.state as dest_state ,
+                          a.departure_time ,
+                          NULL as actual_departure_time ,
+                          NULL as departure_delay ,
+                          NULL as taxi_out ,
+                          NULL as wheels_off ,
+                          NULL as wheels_on ,
+                          NULL as taxi_in ,
+                          a.arrival_time ,
+                          NULL as actual_arrival_time ,
+                          NULL as arrival_delay ,
+                          False as cancelled ,
+                          NULL as cancellation_code ,
+                          a.flight_time ,
+                          NULL as actual_flight_time ,
+                          NULL as air_time ,
+                          1 as flights ,
+                          a.distance ,
+                          NULL as airline_delay ,
+                          NULL as weather_delay ,
+                          NULL as nas_delay ,
+                          NULL as security_delay ,
+                          NULL as late_aircraft_delay ,
+                          -- CONCAT(a.flight_date, '_', a.airline, '_', a.flight_number) as flightDate_airline_flightNumber
+                         FROM {known_args.flights_ext_table} a
+                          LEFT JOIN {known_args.airports_table} b on a.src = b.iata
+                          LEFT JOIN {known_args.airports_table} c on a.dest = c.iata
+                        )"""
+    bq_utils.execute(sql)
 
-    logger.info(f"total time: {(now() - t0):,.6f} secs")
+    logger.info(f"process completed in {(time() - t0):,.3f} seconds")
 
 
 if __name__ == '__main__':
-    run()
-
+    run_simple()
