+++ deb/ch2/ep3/answers/part5.py
