+++ deb/ch2/ep3/answers/part6.py
